backend/pyproject.toml
[project]
name = "smartcompare-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmed",email = "kingzatel@gmail.com"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.128.2,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "openai (>=2.17.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "supabase (>=2.27.3,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.poetry]
package-mode = false